Implement bezier-spline
