+++ packages/turf-bezier-spline/lib/.gitignore
@@ -1,0 +1,1 @@
+spline.js
