+++ packages/turf-bezier-spline/test.js
@@ -1,10 +1,10 @@
-import fs from 'fs';
-import test from 'tape';
-import path from 'path';
-import load from 'load-json-file';
-import write from 'write-json-file';
-import { featureCollection } from '@turf/helpers';
-import bezierSpline from '.';
+const fs = require('fs');
+const test = require('tape');
+const path = require('path');
+const load = require('load-json-file');
+const write = require('write-json-file');
+const { featureCollection } = require('@turf/helpers');
+const bezierSpline = require('.').default;
 
 const directories = {
     in: path.join(__dirname, 'test', 'in') + path.sep,
